refactor: challenges 비즈니스 로직과 렌더링 로직 분리
